# nmrx2/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NMRx 2.0 Research Backend",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/molecules/validate": {
      "post": {
        "summary": "Validate",
        "operationId": "validate_v1_molecules_validate_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Molecule"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/quantum": {
      "post": {
        "summary": "Quantum",
        "operationId": "quantum_v1_jobs_quantum_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quantum"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/docking": {
      "post": {
        "summary": "Docking",
        "operationId": "docking_v1_jobs_docking_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Docking"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{ident}": {
      "get": {
        "summary": "Job",
        "operationId": "job_v1_jobs__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/spectral-fit": {
      "post": {
        "summary": "Fit",
        "operationId": "fit_v1_validation_spectral_fit_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evidence"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/guide": {
      "post": {
        "summary": "Guide",
        "operationId": "guide_v1_guide_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Guide"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Docking": {
        "properties": {
          "smiles": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Smiles"
          },
          "conformers": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Conformers",
            "default": 5
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Seed",
            "default": 42
          },
          "receptor_pdbqt": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 50,
            "title": "Receptor Pdbqt"
          },
          "center_angstrom": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Center Angstrom"
          },
          "box_angstrom": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Box Angstrom",
            "default": [
              20,
              20,
              20
            ]
          },
          "exhaustiveness": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Exhaustiveness",
            "default": 8
          },
          "poses": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Poses",
            "default": 5
          },
          "preparation_notes": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Preparation Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "smiles",
          "receptor_pdbqt",
          "center_angstrom",
          "preparation_notes"
        ],
        "title": "Docking"
      },
      "Evidence": {
        "properties": {
          "observed": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 2,
            "title": "Observed"
          },
          "predicted": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 2,
            "title": "Predicted"
          },
          "sigma": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 2,
            "title": "Sigma"
          },
          "modality": {
            "type": "string",
            "enum": [
              "nmr_ppm",
              "ir_cm-1"
            ],
            "title": "Modality"
          },
          "assignment_provenance": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 5,
            "title": "Assignment Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observed",
          "predicted",
          "sigma",
          "modality",
          "assignment_provenance"
        ],
        "title": "Evidence"
      },
      "Guide": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "use_external_ai": {
            "type": "boolean",
            "title": "Use External Ai",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "Guide"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Molecule": {
        "properties": {
          "smiles": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Smiles"
          },
          "conformers": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Conformers",
            "default": 5
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Seed",
            "default": 42
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "Molecule"
      },
      "Quantum": {
        "properties": {
          "smiles": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Smiles"
          },
          "conformers": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Conformers",
            "default": 5
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Seed",
            "default": 42
          },
          "task": {
            "type": "string",
            "enum": [
              "orbitals",
              "ir",
              "nmr"
            ],
            "title": "Task",
            "default": "orbitals"
          },
          "method": {
            "type": "string",
            "enum": [
              "HF",
              "B3LYP",
              "PBE0"
            ],
            "title": "Method",
            "default": "B3LYP"
          },
          "basis": {
            "type": "string",
            "enum": [
              "sto-3g",
              "def2-svp",
              "def2-tzvp"
            ],
            "title": "Basis",
            "default": "def2-svp"
          },
          "optimize": {
            "type": "boolean",
            "title": "Optimize",
            "default": true
          },
          "displacement_bohr": {
            "type": "number",
            "maximum": 0.02,
            "minimum": 0.001,
            "title": "Displacement Bohr",
            "default": 0.005
          },
          "reference_shielding_ppm": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Reference Shielding Ppm"
          },
          "reference_provenance": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Provenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "Quantum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}